Implemented init() method in Anonymous module.
Following my proposal on MESOS-4253[0] I have
implemented the init() module to gather the Agent's flags
(the demo shows how the module can get access to the
work_dir and sandbox_dir flags).

Also updated the build to fix an error (the module shared
library should *not* link with libmesos).

Finally, some minor refactoring to rename the methods in a way
that more closely reflects the API.