Refactored to use mesos::CommandInfo
We prefer to use already-defined protocol buffers, instead of
crafting our own; in addition, `CommandInfo` provides a few other
functionalities (environment variables, URIs) that may prove to be
useful in future.

Also taken the opportunity to clean up some of the code style,
and update some of the tests.

The README has been updated with the change of API and the version
bumped up to 0.3.0.